pyproject.toml
[tool.poetry]
name = "hh-parser"
version = "0.1.0"
description = ""
authors = ["Andrew <orwellj@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
